sept submission and sept report is done

+++ Swdn-service/src/main/java/com/swdn/model/response/SeptSubmissionResponse.java
@@ -15,7 +15,7 @@ public class SeptSubmissionResponse {
 
 	private String septProfile;
 
-	private Integer netResultPercentage;
+	private float netResultPercentage;
 
 	private String resultStatement;
